:smiley_cat: Google auth client, gene T2T-CHM13 locations/orientations, DNA sequence search

+++ src/main/java/org/magicat/model/SequenceItem.java
@@ -24,7 +24,7 @@ public class SequenceItem {
     private List<String> name;
 
     @Field("position")
-    private List<Integer> position;
+    private List<Long> position;
 
     @Field("genbank")
     private List<String> genbank;
@@ -63,11 +63,11 @@ public class SequenceItem {
         this.name = name;
     }
 
-    public List<Integer> getPosition() {
+    public List<Long> getPosition() {
         return position;
     }
 
-    public void setPosition(List<Integer> position) {
+    public void setPosition(List<Long> position) {
         this.position = position;
     }
